requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

ultralytics
kaggle
pandas
numpy
opencv-python
matplotlib
ipython
pyyaml
torch
torchvision
streamlit
ultralytics
opencv-python-headless
pillow